DTU: the registers have to be 64 bits large.
Because some of them contain pointers.

+++ src/mem/dtu/regfile.hh
@@ -73,7 +73,7 @@ class RegFile
 {
   public:
 
-    using reg_t = uint32_t;
+    using reg_t = uint64_t;
 
     static Addr getRegAddr(DtuReg reg);
